# openapi.yaml
openapi: 3.0.0
info:
  title: OVS 管理平台 API
  version: 1.0.0
  description: |-
    Open vSwitch 管理平台后端 API，涵盖交换机、端口、Bond、流表、镜像、QoS、隧道、BFD、CFM、组播、STP/RSTP、NetFlow/sFlow/IPFIX、网络命名空间等全部常用与高级功能。
servers:
  - url: http://localhost:8080
paths:
...
  /api/ovs/mirror/add:
    post:
      summary: 新增端口镜像
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                name:
                  type: string
                selectSrcPorts:
                  type: array
                  items:
                    type: string
                selectDstPorts:
                  type: array
                  items:
                    type: string
                selectVlan:
                  type: integer
                outputPort:
                  type: string
                outputVlan:
                  type: integer
                selectAll:
                  type: boolean
              required: [bridge, name]
      responses:
        '200':
          description: 成功
  /api/ovs/mirror/delete:
    post:
      summary: 删除端口镜像
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
                name:
                  type: string
              required: [bridge, name]
      responses:
        '200':
          description: 成功
  /api/ovs/mirror/list:
    post:
      summary: 查询端口镜像
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bridge:
                  type: string
              required: [bridge]
      responses:
        '200':
          description: 成功
...
  /api/netns/create:
    post:
      summary: 新增命名空间
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        '200':
          description: 成功
  /api/netns/delete:
    post:
      summary: 删除命名空间
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        '200':
          description: 成功
  /api/netns/list:
    post:
      summary: 查询命名空间列表
      requestBody:
        required: false
      responses:
        '200':
          description: 成功 